gui/requirements.txt
# Video Downloader GUI - Requirements
# Python 3.8+ required

# Core video download functionality
yt-dlp>=2024.12.0

# GUI dependencies (Tkinter is included with Python)
# No additional GUI packages needed for Tkinter

# Image processing for thumbnails
Pillow>=10.0.0

# Audio metadata handling
mutagen>=1.47.0

# Clipboard functionality
pyperclip>=1.8.0

# HTTP requests
requests>=2.28.0

# CLI version dependencies (for compatibility)
rich>=13.0.0

# Optional: Additional audio/video codecs
# FFmpeg should be installed separately on the system